+++ cuda/lltorque.go
@@ -25,7 +25,7 @@ func LLTorque(torque, m, B *data.Slice, alpha MSlice) {
 		k_lltorque2time_async(torque.DevPtr(X), torque.DevPtr(Y), torque.DevPtr(Z),
 			m.DevPtr(X), m.DevPtr(Y), m.DevPtr(Z),
 			B.DevPtr(X), B.DevPtr(Y), B.DevPtr(Z),
-			alpha.DevPtr(0), alpha.Mul(0), float32(Time_cuda), Brms_cuda[X], Brms_cuda[Y], Brms_cuda[Z],
+			alpha.DevPtr(0), alpha.Mul(0), float32(Time_cuda), float32(Brms_cuda[X]), float32(Brms_cuda[Y]), float32(Brms_cuda[Z]),
 			M_rk.DevPtr(0), M_rk.DevPtr(1), M_rk.DevPtr(2), M_rk.DevPtr(3), N, cfg)
 
 	} else {
